+++ libyasm/phash.h
@@ -10,6 +10,6 @@ Source is http://burtleburtle.net/bob/c/lookupa.h
 ------------------------------------------------------------------------------
 */
 
-unsigned long phash_lookup(const char *k, unsigned long length,
+unsigned long phash_lookup(const char *k, size_t length,
 			   unsigned long level);
-void phash_checksum(const char *k, unsigned long length, unsigned long *state);
+void phash_checksum(const char *k, size_t length, unsigned long *state);
